template-matching/requirements.txt
pandas>=0.25,<3
streamlit==1.24.1
langchain==0.0.340
faiss-cpu==1.7.4
openai==1.3.5
powerbiclient==3.1.1
spacy==3.7.2
gensim==4.3.2
numpy==1.25.1
tiktoken==0.5.1
pillow>=8.4.0,<10.2.0